add globals_ file
