+++ flare/components/base.py
@@ -9,7 +9,7 @@ import hikari
 import sigparse
 
 from flare.exceptions import MissingRequiredParameterError, SerializerError
-from flare.internal import event_handler
+from flare.internal import globals_
 
 if t.TYPE_CHECKING:
     from flare import context
@@ -39,9 +39,9 @@ class Component(abc.ABC, t.Generic[P]):
 
         if not self.args:
             # If no args were passed, calling set() isn't necessary to construct custom_id
-            self._custom_id = event_handler.active_serde.serialize(self.cookie, {}, {})
+            self._custom_id = globals_.active_serde.serialize(self.cookie, {}, {})
 
-        event_handler.components[self.cookie] = self
+        globals_.components[self.cookie] = self
 
     @property
     @abc.abstractmethod
@@ -90,16 +90,14 @@ class Component(abc.ABC, t.Generic[P]):
         assert component.custom_id
 
         try:
-            flare_component, kwargs = event_handler.active_serde.deserialize(
-                component.custom_id, event_handler.components
-            )
+            flare_component, kwargs = globals_.active_serde.deserialize(component.custom_id, globals_.components)
         except SerializerError:
             raise
         return flare_component.set(kwargs)
 
     def set(self: ComponentT, *_: P.args, **values: P.kwargs) -> ComponentT:
         new = copy.copy(self)  # Create new instance with params set
-        new._custom_id = event_handler.active_serde.serialize(self.cookie, self.args, values)
+        new._custom_id = globals_.active_serde.serialize(self.cookie, self.args, values)
         return new
 
     def as_keyword(self, args: list[t.Any], kwargs: dict[str, t.Any]) -> dict[str, t.Any]:
