+++ flare/context.py
@@ -5,6 +5,7 @@ import typing as t
 
 import hikari
 from hikari.snowflakes import Snowflake
+
 from flare import row
 
 __all__: t.Sequence[str] = ("Context", "InteractionResponse")
@@ -235,7 +236,7 @@ class Context:
         return self._interaction.get_channel()
 
     def get_components(self) -> t.MutableSequence[row.Row]:
-        """Returns the flare components for the interaction this context is proxying."""
+        """Returns the flare components for the interaction this context is proxying"""
         return row.Row.from_message(self.message)
 
     async def get_last_response(self) -> InteractionResponse:
