+++ flare/internal/README.md
@@ -3,6 +3,9 @@
 This directory has the tools used to interact with the api. These should not be touched
 by the end user.
 
+## attrs
+Contains global information for `flare`.
+
 ## event_handler
 
 The file containing the callback called when a interaction is received.
