+++ flare/internal/event_handler.py
@@ -1,41 +1,15 @@
 import logging
-import typing as t
 
 import hikari
 
 from flare.context import Context
 from flare.exceptions import SerializerError
-from flare.internal.serde import Serde, SerdeABC
-
-__all__: t.Final[t.Sequence[str]] = ("install",)
+from flare.internal import globals_
 
 logger = logging.getLogger(__name__)
 
-components: dict[str, t.Any] = {}
-"""Currently loaded components."""
 
-active_serde: SerdeABC = Serde()
-"""The currently active serializer."""
-
-
-def install(bot: hikari.EventManagerAware, serde: SerdeABC | None = None) -> None:
-    """Install flare under the given bot instance.
-
-    Args:
-        bot:
-            The bot to install flare under.
-        serde:
-            For advanced usage, you can pass a custom serializer. By default uses the default serializer.
-    """
-    global active_serde
-
-    if serde is not None:
-        active_serde = serde
-
-    bot.event_manager.subscribe(hikari.InteractionCreateEvent, _on_inter)
-
-
-async def _on_inter(event: hikari.InteractionCreateEvent) -> None:
+async def on_inter(event: hikari.InteractionCreateEvent) -> None:
     """
     Function called to respond to an interaction.
     """
@@ -43,7 +17,7 @@ async def _on_inter(event: hikari.InteractionCreateEvent) -> None:
         return
 
     try:
-        component, kwargs = active_serde.deserialize(event.interaction.custom_id, components)
+        component, kwargs = globals_.active_serde.deserialize(event.interaction.custom_id, globals_.components)
     except SerializerError:  # If the custom_id is invalid, it was probably not created by flare.
         logger.debug(
             f"Flare received custom_id '{event.interaction.custom_id}' which it cannot deserialize.", exc_info=True
